Refactor code to multiple files, use os environment variables for Ginkgo test inputs
